VILLAGER-TYPE : Implemented db dao for VillagerType
* Upgraded to mc-data version 1.1.4 to get public constructors
* Reorganized test stub programs.  One package for each entity

NOTE: I tried setting the parent type of the VillagerType entity in the persistence
configuration file.  It accepted this, but it did not recognize the db.entity type
as the preferred implementation.  Because of this, I did not get the table built
for me in the database.  This is another area where eventual convergence will make
these little wrinkles go away.